requirements.txt
flask

sqlalchemy
python-dotenv
requests
